exp/precomp/requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2019.11.28
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.18.5
opencv-python==4.1.2.30
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pbr==5.4.4
pdf2image==1.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2018.11.26
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.4
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
stevedore==1.32.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
tornado==6.0.4
traitlets==4.3.3
virtualenv==20.0.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.0.0
